# Perform BitStream I/O operations on underlying buffers