It seems to be quite a bit faster to perform I/O directly on the stream
buffer returned by rdbuf() than to go through the stream's read/write
methods. I'm guessing this is due to reduced virtual method invocations.

Coding times on a 100MB file decreased as follows:

    Encoding: 0.65s -> 0.54s
    Decoding: 1.10s -> 0.97s